Rename connect4 strategy
